pyproject.toml
[[tool.poetry.source]]
name = "default"
url = "https://pypi.org/simple"

[tool.poetry]
name = "pymft"
version = "0.1.1"
description = "Library for Midi Fighter Twister Interfacing in Python"
authors = ["Sina Solaimanpour <sinas.cb@gmail.com>"]


[tool.poetry.dependencies]
python = ">=3.10,<4.0"
typing = "^3.7.4"
python-rtmidi = "^1.5.8" # https://github.com/SpotlightKid/python-rtmidi


[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
pytest-mock = "^3.6.1"
mypy = "^1.5.1"
black = "^22.3.0"
usort = "^0.6.4"
isort = {extras = ["colors"], version = "^5.9.3"}

[tool.poetry.scripts]
demo_main = "pymft.main:run"
version = "pymft.main:version"

[tool.mypy]
plugins = []
exclude = []

[tool.black]
line-length = 80
target-version = ["py311"]

[tool.isort]
profile = "black"
known_first_party = ["pymft"]
color_output = true
line_length = 80

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"